Update nf-synchapi-initonceexecuteonce.md
I felt the existing text did not describe the full operation, in particular with respect to the different behavior between the callback returning TRUE vs FALSE.  As well, the example code linked in these pages doesn't really illustrate the nuances of this operation.  I have attempted to explain the success and failure cases separately and in more detail.